Consumable/Trinket functionality added, art assets imported

+++ 2D Game Project (2019)/Assets/Scripts/Characters/Base/Character2D.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 public abstract class Character2D : MonoBehaviour
 {
     // Health
-    [SerializeField] protected int maximumHealth = 100;
-    public int MaximumHealth { get { return maximumHealth; } set { maximumHealth = value; } }
-    protected int currentHealth;
-    public int CurrentHealth { get { return currentHealth; } set { currentHealth = value; } }
+    [SerializeField] protected float maximumHealth = 100;
+    public float MaximumHealth { get { return maximumHealth; } set { maximumHealth = value; } }
+    protected float currentHealth;
+    public float CurrentHealth { get { return currentHealth; } set { currentHealth = value; } }
 
     // Movement
     [SerializeField] protected int speed = 5;
